# Use getTopicTypes() in /rosapi/get_topics for increased performance