Hello,

When using `rosapi` in the development of web applications for the PR2 robot, I was facing the problem that the call for the `/rosapi/topics` service was extremely slow (operating with over 1200 topics).

Looking into the problem, it seems that getting the types was taking a lot of time:

```python
In [3]: time topics = get_topics(['*'])
CPU times: user 152 ms, sys: 0 ns, total: 152 ms
Wall time: 183 ms

In [4]: time types = get_topics_types(topics, ['*'])
CPU times: user 1min 18s, sys: 428 ms, total: 1min 18s
Wall time: 2min 6s
```

By using the `Master.getTopicTypes()` we can get equivalent results (the order changes) with better performance:
```python
In [5]: time new_topics, new_types = get_topics_and_types(['*'])
CPU times: user 244 ms, sys: 0 ns, total: 244 ms
Wall time: 297 ms

In [6]: set(new_topics) == set(topics) and set(types) == set(new_types)
Out[6]: True
```

Note that differently from `Master.getSystemState()`, `Master.getTopicTypes()` also includes inactive topics and does not include topics with unknown messages.

Related to: https://github.com/ros/ros_comm/issues/53